+++ lib/puffer_pages/backends/models/layout.rb
@@ -3,4 +3,8 @@ class PufferPages::Layout < ActiveRecord::Base
 
   validates_presence_of :name
   validates_uniqueness_of :name
+
+  def source
+    body
+  end
 end