+++ lib/puffer_pages/backends/models/page.rb
@@ -34,7 +34,12 @@ class PufferPages::Page < ActiveRecord::Base
     map{|page| page.to_drop(*args)}
   end
 
-  has_many :page_parts, :order => "name = '#{PufferPages.primary_page_part_name}' desc, name", :dependent => :destroy, :class_name => '::PagePart', :validate => true
+  has_many :page_parts,
+    :order => "name = '#{PufferPages.primary_page_part_name}' desc, name",
+    :dependent => :destroy,
+    :class_name => '::PagePart',
+    :validate => true,
+    :inverse_of => :page
   accepts_nested_attributes_for :page_parts, :allow_destroy => true
   belongs_to :layout, :primary_key => :name, :foreign_key => :layout_name, :class_name => '::Layout'
 
@@ -75,16 +80,19 @@ class PufferPages::Page < ActiveRecord::Base
     PufferPages.single_section_page_path ? slug : location
   end
 
-  def render drops_or_context = {}
+  def render context = {}
     if inherited_layout
-      render_layout(inherited_layout.body, drops_or_context)
+      render_layout(inherited_layout.body, context)
     else
-      inherited_page_parts.reverse.map{|part| part.render(drops_or_context, self)}.join
+      inherited_page_parts.reverse.map{|part|
+        result = part.render(context, self)
+        part.main? ? result : "<% content_for :'#{part.name}' do %>#{result}<% end %>"
+      }.join
     end
   end
 
-  def render_layout layout, drops_or_context = {}
-    render_liquid(layout, self, drops_or_context)
+  def render_layout layout, context = {}
+    render_liquid(layout, self, context)
   end
 
   def inherited_layout_page
@@ -109,6 +117,10 @@ class PufferPages::Page < ActiveRecord::Base
 
   def all_inherited_page_parts
     @all_inherited_page_parts ||= ::PagePart.where(:page_parts => {:page_id => self_and_ancestors.map(&:id)}).joins(:page).order("page_parts.name = '#{PufferPages.primary_page_part_name}' desc, page_parts.name, pages.lft desc")
+  end
+
+  def super_inherited_page_parts
+    @super_inherited_page_parts ||= ::PagePart.where(:page_parts => {:page_id => ancestors.map(&:id)}).joins(:page).order("page_parts.name = '#{PufferPages.primary_page_part_name}' desc, page_parts.name, pages.lft desc")
   end
 
   def part name